Re-vendor OBML converter: split monolith into focused modules
Sync the vendored converter with the upstream OrionBelt package, which was
decomposed from a single 2967-line converter.py into a thin facade plus
focused modules. Public API (osi_orionbelt.converter.*) and behavior are
unchanged; the converter test suite passes (141).

- converter.py: now a facade re-exporting every public name + main()
- _common.py: shared constants / type maps
- osi_to_obml.py: OSItoOBML (OSI -> OBML)
- obml_to_osi.py: OBMLtoOSI (OBML -> OSI)
- ontology.py: OBMLtoOSIOntology
- validation.py: ValidationResult + validate_obml/osi/osi_ontology
